Pin Dockerfile base image to Alpine 3.20
Both production and development environments are using Alpine 3.20. for `crossreference-app`.
The version is being pinned because Alpine 3.23 includes newer MariaDB packages specific to the "11.4.x" series,
which causes a client SSL issue during connections, both from the command-line client and the mariadb python module.
```
    /app/src # mariadb -h 127.0.0.1 -P 3306 -u buildmaster -p
    Enter password:
    ERROR 2026 (HY000): TLS/SSL error: SSL is required, but the server does not support it
```

This issue has not been investigated in detail, as it is outside the scope of this patch.
Upgrades should be handled and tested separately (the app requires a rewrite from Django 3 to Django 5).

For now, pinning the base image version ensures the application continues to work correctly.

The next commit will add the DRF feature.