Expose CrossReference data via a REST API
To make CrossReference data usable in scripts, applications, and dashboards
(e.g., the Release dashboard), this patch exposes the data through a REST API.

This patch implements the route `cr/api/testfailures/` which provides read-only access
to the data via a `ViewSet` (i.e., `ReadOnlyModelViewSet`).

It reuses the existing `TestFailure` model (`class TestFailure(models.Model)`)
that is also used to display data on the `/cr` web page.

The `select_test_failures` function returns a `QuerySet` that is
then serialized to JSON and returned to the client.
Query parameters are validated using the `APIQueryParamsSerializer`.